src/yaaos-sfs/pyproject.toml
[project]
name = "yaaos-sfs"
version = "0.2.0"
description = "YAAOS Semantic File System — search files by meaning, not name"
requires-python = ">=3.12"
dependencies = [
    "sentence-transformers>=3.0",
    "sqlite-vec>=0.1",
    "watchdog>=4.0",
    "click>=8.0",
    "rich>=13.0",
    "tomli>=2.0",
    "pymupdf>=1.24",
    "pathspec>=0.12.0",
    "xxhash>=3.4.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
voyage = ["voyageai>=0.3"]
docs = [
    "python-docx>=1.1",
    "python-pptx>=1.0",
    "openpyxl>=3.1",
    "ebooklib>=0.18",
    "striprtf>=0.0.26",
]
media = [
    "Pillow>=10.0",
    "mutagen>=1.47",
]
code = [
    "tree-sitter-languages>=1.10",
]
all = [
    "yaaos-sfs[docs,media,code]",
]
dev = ["pytest>=8.0", "ruff>=0.8"]

[project.scripts]
yaaos-find = "yaaos_sfs.cli:main"
yaaos-sfs = "yaaos_sfs.daemon:main"

[dependency-groups]
dev = ["pytest>=8.0", "ruff>=0.8"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
